Move initializeDependencies in its own integration module
Summary:
This was already used at several places outside of the main application,
including tests, it was time to move it out. Also it will be used in
further places soon. Last but not least, I renamed the only fonction to
make it clearer that it initializes the global dependency manager for
use in the application.

Reviewers: dfaure

Differential Revision: https://phabricator.kde.org/D23887