Port all of inboxpagemodeltest away from mocks
Summary:
* Had to extend error handling in akonadifakestorage

* There is something nasty with the kjobs being triggered for extra
info, from *PageModel::data(). If they keep running past the end
of the test method, they end up calling into a deleted Domain::LiveQueryInput.
This is because JobHandlerInstance is a singleton so it stays around,
the job's signal calls the JobHandlerInstance slot, and from there it's all
about lambdas in deleted objects.
Calling waitForEmptyJobQueue() at the right places avoids this, but it's
still unexpected.

* I wanted to skip fetching extra info when role == ObjectRole
(we don't need it for that), but this only triggers the above problem
even more, now it's later calls to data() which fetch extra info,
increasing chances of jobs lying around after deletion of everything
else.

Test Plan: Tests pass

Reviewers: ervin

Reviewed By: ervin

Differential Revision: https://phabricator.kde.org/D24282