Disable running task widget on Wayland
There are two problems with it:

1) RunningTaskWidget is a plain QWidget that gets turned into a window by calling winId()
On Wayland that causes a subsurface to be created, which doesn't work as expected and causes interaction with
the whole app to be severely broken on Wayland. The fix for this would be to use a proper window class
like QDialog as base.

2) Even if the above issues was fixed we still wouldn't get the result we want. On Wayland normal application
windows can't absolutely postion themselves on the screen like we do here to adhere to the top screen edge.
This could potentially be fixed by using the layer-shell wayland protocol.

Disable the running task widget to make Zanshin usable on Wayland, albeit with slightly less features

(cherry picked from commit 8d6d74240bd35882517d2e6282c44ce3cb7bf3d3)