Fix doubled projects after unchecking and checking again a data source
Calling collectionChanged() and collectionSelectionChanged() immediately
afterwards leads to the following problem:

- collectionChanged() => LiveQuery::result => doFetch => LiveQueryHelpers::fetchItems
  => CachingCollectionItemsFetchJob::start => delayed call to retrieveFromCache
- collectionSelectionChanged() => LiveQuery::reset => clear() and
  doFetch() => fetchItems etc. same as above.

So when clear() is called, there's nothing to clear yet (fetching being
delayed). And then both fetch jobs run and append all items twice.

This commit ensures that the first job is killed when starting the
second one for the same collection.